[refactor] refactor in marketplace page create marketplace layout
