+++ src/pages/Marketplace.tsx
@@ -1,26 +1,17 @@
-import Breadcrumb from '../component/Breadcrumb';
 import { CardProduct } from '../component/Cards';
-import { Navigation, NavigationMarketplace } from '../component/Navigations';
+import MarketplaceLayout from '../component/layouts/MarketplaceLayout';
 
 export default function Marketplace() {
   return (
-    <div className="w-full min-h-dvh">
-      <Navigation />
-      <NavigationMarketplace />
-      <div className="max-w-screen-xl mx-auto my-7">
-        <Breadcrumb />
-        <div className="grid grid-cols-5 gap-4 mt-6">
-          <CardProduct title="Title Product" price={100000} discountPersent={10} rating={3} reviews={5} shop="Name Shop" />
-          <CardProduct title="Title Product" price={100000} discountPersent={10} rating={3} reviews={5} shop="Name Shop" />
-          <CardProduct title="Title Product" price={100000} discountPersent={10} rating={3} reviews={5} shop="Name Shop" />
-          <CardProduct title="Title Product" price={100000} discountPersent={10} rating={3} reviews={5} shop="Name Shop" />
-          <CardProduct title="Title Product" price={100000} discountPersent={10} rating={3} reviews={5} shop="Name Shop" />
-          <CardProduct title="Title Product" price={100000} discountPersent={10} rating={3} reviews={5} shop="Name Shop" />
-          <CardProduct title="Title Product" price={100000} discountPersent={10} rating={3} reviews={5} shop="Name Shop" />
-          <CardProduct title="Title Product" price={100000} discountPersent={10} rating={3} reviews={5} shop="Name Shop" />
-          <CardProduct title="Title Product" price={100000} discountPersent={10} rating={3} reviews={5} shop="Name Shop" />
-        </div>
+    <MarketplaceLayout>
+      <div className="grid grid-cols-5 gap-4 mt-6">
+        <CardProduct image="https://placehold.co/800@3x.png" title="Title Product" price={100000} discountPersent={10} rating={3} reviews={5} shop="Name Shop" />
+        <CardProduct image="https://placehold.co/800@3x.png" title="Title Product" price={100000} discountPersent={10} rating={3} reviews={5} shop="Name Shop" />
+        <CardProduct image="https://placehold.co/600x400@2x.png" title="Title Product" price={100000} discountPersent={10} rating={3} reviews={5} shop="Name Shop" />
+        <CardProduct image="https://placehold.co/600x400@2x.png" title="Title Product" price={100000} discountPersent={10} rating={3} reviews={5} shop="Name Shop" />
+        <CardProduct image="https://placehold.co/800@3x.png" title="Title Product" price={108000} discountPersent={10} rating={3} reviews={5} shop="Name Shop" />
+        <CardProduct image="https://placehold.co/600x400@2x.png" title="Title Product" price={108000} discountPersent={10} rating={3} reviews={5} shop="Name Shop" />
       </div>
-    </div>
+    </MarketplaceLayout>
   );
 }
